(partially working) move facility summary functions to shared utils file
