+++ src/page-weekly-snapshot/SystemWideSummaryTable.tsx
@@ -6,17 +6,25 @@ import {
   findMatchingDay,
   findMostRecentDate,
 } from "../hooks/useAddCasesInputs";
+import { Facility } from "../page-multi-facility/types";
 import { Table } from "./FacilityPage";
 import { HorizontalRule, LeftHeading } from "./shared/index";
 import {
   buildIncarceratedFacilitySummaryData,
+  IncarceratedFacilitySummaryData,
   makeTableHeadings,
 } from "./shared/utils";
 
-const SystemWideSummaryTable: React.FC<{}> = () => {
-  const { state: facilitiesState } = useFacilities();
+function getSystemWideSummaryIncarceratedData(facilities: Facility[]) {
+  let systemWideSummaryIncarceratedData: IncarceratedFacilitySummaryData = {
+    incarceratedPopulation: 0,
+    incarceratedPopulationDelta: 0,
+    incarceratedPopulationDeltaDirection: "",
+    incarceratedCases: 0,
+    incarceratedCasesDelta: 0,
+    incarceratedCasesDeltaDirection: "",
+  };
 
-  const facilities = Object.values(facilitiesState.facilities);
   for (let i = 0; i < facilities.length; i++) {
     const facility = facilities[i];
     const hasEarlierData = facility.modelVersions.length > 1;
@@ -38,7 +46,36 @@ const SystemWideSummaryTable: React.FC<{}> = () => {
       facility,
       facility.modelInputs,
     );
+    console.log(incarceratedData);
+
+    systemWideSummaryIncarceratedData.incarceratedPopulation +=
+      incarceratedData.incarceratedPopulation;
+    // if (incarceratedData.incarceratedPopulationDeltaDirection == "negative") {
+    //     systemWideSummaryIncarceratedData.incarceratedPopulationDelta -= incarceratedData.incarceratedPopulationDelta;
+    // }
+    // else {
+    //     systemWideSummaryIncarceratedData.incarceratedPopulationDelta += incarceratedData.incarceratedPopulationDelta;
+    // }
+    systemWideSummaryIncarceratedData.incarceratedCases +=
+      incarceratedData.incarceratedCases;
+    // if (incarceratedData.incarceratedCasesDeltaDirection == "negative") {
+    //     systemWideSummaryIncarceratedData.incarceratedCasesDelta -= incarceratedData.incarceratedCasesDelta;
+    // }
+    // else {
+    //     systemWideSummaryIncarceratedData.incarceratedCasesDelta += incarceratedData.incarceratedCasesDelta;
+    // }
   }
+  return systemWideSummaryIncarceratedData;
+}
+
+const SystemWideSummaryTable: React.FC<{}> = () => {
+  const { state: facilitiesState } = useFacilities();
+
+  const facilities = Object.values(facilitiesState.facilities);
+  const systemWideSummaryIncarceratedData = getSystemWideSummaryIncarceratedData(
+    facilities,
+  );
+
   return (
     <>
       <HorizontalRule />
